wip: start refactor for implement plugin system
